Refactor: Move taker fee and maker fee methods to TradeFeeValidation
