+++ core/src/main/java/bisq/core/trade/bsq_swap/BsqSwapTakeOfferRequestVerification.java
@@ -37,8 +37,8 @@ import java.util.concurrent.TimeUnit;
 
 import lombok.extern.slf4j.Slf4j;
 
-import static bisq.core.trade.validation.TradeValidation.checkMakerFee;
-import static bisq.core.trade.validation.TradeValidation.checkTakerFee;
+import static bisq.core.trade.validation.TradeFeeValidation.checkMakerFee;
+import static bisq.core.trade.validation.TradeFeeValidation.checkTakerFee;
 import static bisq.core.trade.validation.TradeValidation.checkTradeId;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
